fixed a bug in the calculation of expression domain for ConstantExpressions:
In fact, { [] : } is not the same as { : }. One is zero-dimensional set,
but the other is not even a set (the in dim does not exist). The buggy
implementation used buildUniverse on the parameter space, but this was
not creating zero-dim sets. Now it generates the appropriate string to
construct a scalar dim explicitly from strings.