fixed issues in Alpha visitors:
There were some issues due to how classes with inheritance were
interacting with the visitor. When XXX is an abstract class,
 visitXXX was either never called or caused inXXX and inYYY
for both XXX and its subclass YYY to be both called in a single
traversal (same for outXXX and outYYY).

The model was modified so that visitXXX is defined in the super class
if the underlying structure is identical (e.g., AbstractReduceExpression).
The appropriate in/out method is then called by sequence of instanceof tests.
When the structure is different, then the visitXXX was removed from the
abstract class, and are only invoked in the concrete implementations.
This has caused some classes to become interfaces.