giant commit related to adding PolynomialIndexExpression:
  - added PolynomialIndexExpression; a variant of IndexExpression that
takes PWQPolynomial in ISL instead of affine functions.
  - added JNIPolynomial and JNIPolynomialInArrayNotation to represent
polynomials. These two are part of CalculatorExpression, but are only
used within PolynomialIndexExpression.
  - The basic chain for AlphaInternalStateConstructor, PrintAST, Show,
and AShow are updated to handle PolynomialIndexExpression
  - The Xtext parser is updated to handle PolynomialIndexExpression
  - Fixed bug in parser that made IndexExpresssion with array notation
omit the keyword 'val'
  - Fixed issues with validator not checking duplicate names accross
files anymore
  - Enhanced error reporting for duplicate names by providing the list
of files containing the duplicates. Since the check is expensive, it is
not ran all the time, and hence duplicates are only flagged for files
under edit. This message should make it easier to spot where the other
duplicates are.
  - Other minor tweaks on label printing etc.