Fixing the dependence raising rule for restrictions
The previous iteration of the rule appeared to be the inverse of the
associated normalization rule, but was not. Instead, information about
the restriction domain was being lost in some cases, resulting in
incorrect programs. This was most easily seen when applying dependence
raising to a prefix scan equation. A test for this was added to ensure
this scenario doesn't occur again.