Codegen: adding unit tests and fixes for function building
If there are multiple outputs, then the entry point for the system may
use the same loop variables multiple times (for each output variable).
To work around this, the function builder now checks if a variable
declaration is a duplicate and skips it if so (or throws an exception if
it's the same name but a different type).