Factorizing affine expressions now only factorizes the linear part
The parameters/constants are stripped at the beginning, then added back
at the end. Only the linear part of the affine expressions are factored,
that way they're never produced in the inner dimensions.

Also fixed how the maps were created if the common factor was a map from
inputs to the empty set (or only parameters/constants).